Always encrypt Client ID
This is what the official CDM does on platforms with Verified Media
Path, and it seems beneficial for privacy regardless. If we aren't given
a service certificate, we can request one from the license server prior
to requesting the license itself. Refactor Session to be more of a
proper state machine so that we can accommodate the extra round trip.