Use separate manifests for Firefox and Chrome
We cannot tell Firefox that we support any codecs, as it will
unconditionally use DecryptAndDecodeFrame(), which we don't implement at
all, for such codecs with no fallback to Decrypt(). Conversely, we must
tell Chrome that we support all codecs, as it won't even instantiate the
CDM for unsupported codecs but calls Decrypt() just fine for supported
ones.