Support interfaces 10 and 11 simultaneously
Since autocxx subclasses don't support multiple inheritance, we have to
subclass all the ContentDecryptionModule_NN versions in C++ and then
subclass a second time in Rust. This means we do now have some C++ in
our tree (hence the Chromium-derived .clang-format), but it's cleaner
than the other options. For Host_NN, we can do dynamic dispatch in Rust.
I've added a macro to reduce the boilerplate, but there's still some.
Even though the code's not the cleanest, I think this feature is
important enough to justify the added complexity.

The new C++ header exposed a bug in autocxx, which I've fixed in [1] and
incorporated into our fork.

Note that, while Chromium parses the "x-cdm-*-versions" manifest keys as
comma-separated lists, Firefox parses each as a single number. I've put
the comma-separated versions in the Firefox manifest too because the bug
might get fixed by the time Firefox drops interface 10 support.

[1] https://github.com/google/autocxx/pull/1486